Updated for Lead Developer position
- Stopped using 'proper' publications.bib approach to managing
links to articles etc. It was too restrictive and I just didn't
like how it formatted the publications. I now manage the styles
directly in template.text.

- Moves Publications into the sidebar, and removes the Key
Achievements. The former was making the doc too long (on top of
job evidence) and the latter was nice but it repeated stuff that
was elsewhere in the CV, so was a waste of valuable real estate.
I also have to let go of my "Best Contributor of 2015" award
eventually...!

- Swaps around the 'location' and 'job title' rendering so the
location comes first. This meant I could remove the repeated
employer names and use the first job title as a kind of heading
that groups all the job titles, saving valuable space.

- Adds the new Lead Developer position, and fills out some of the
Senior Developer-y things I've been doing. Ironically this is the
bit I've spent the least time on, so there may be better examples
of things I can come back to in a future iteration.